# openapi.json
{"openapi":"3.1.0","info":{"title":"TRR Assembly Work Orders API","version":"0.1.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/bootstrap/admin":{"post":{"summary":"Bootstrap First Admin","operationId":"bootstrap_first_admin_bootstrap_admin_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/login":{"post":{"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users":{"get":{"summary":"List Users","operationId":"list_users_users_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserOut"},"title":"Response List Users Users Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"summary":"Create User","operationId":"create_user_users_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/users/reset-pin":{"post":{"summary":"Admin Reset User Pin","operationId":"admin_reset_user_pin_admin_users_reset_pin_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetUserPinRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stations":{"get":{"summary":"Stations","operationId":"stations_stations_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/work-orders":{"post":{"summary":"Create Work Order","operationId":"create_work_order_work_orders_post","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateWORequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WOOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Work Orders","operationId":"list_work_orders_work_orders_get","parameters":[{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WOOut"},"title":"Response List Work Orders Work Orders Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/work-orders/{wo_id}":{"get":{"summary":"Get Work Order","operationId":"get_work_order_work_orders__wo_id__get","parameters":[{"name":"wo_id","in":"path","required":true,"schema":{"type":"integer","title":"Wo Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WOOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/work-orders/{wo_id}/close":{"post":{"summary":"Close Work Order","operationId":"close_work_order_work_orders__wo_id__close_post","parameters":[{"name":"wo_id","in":"path","required":true,"schema":{"type":"integer","title":"Wo Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloseWOResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/work-orders/{wo_id}/notes":{"post":{"summary":"Add Note","operationId":"add_note_work_orders__wo_id__notes_post","parameters":[{"name":"wo_id","in":"path","required":true,"schema":{"type":"integer","title":"Wo Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddNoteRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Notes","operationId":"list_notes_work_orders__wo_id__notes_get","parameters":[{"name":"wo_id","in":"path","required":true,"schema":{"type":"integer","title":"Wo Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NoteOut"},"title":"Response List Notes Work Orders  Wo Id  Notes Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/work-orders/{wo_id}/workers":{"get":{"summary":"List Workers","operationId":"list_workers_work_orders__wo_id__workers_get","parameters":[{"name":"wo_id","in":"path","required":true,"schema":{"type":"integer","title":"Wo Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkerOut"},"title":"Response List Workers Work Orders  Wo Id  Workers Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/work-orders/{wo_id}/workers/start":{"post":{"summary":"Start Working On Wo","operationId":"start_working_on_wo_work_orders__wo_id__workers_start_post","parameters":[{"name":"wo_id","in":"path","required":true,"schema":{"type":"integer","title":"Wo Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/WorkerOut"},"title":"Response Start Working On Wo Work Orders  Wo Id  Workers Start Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/work-orders/{wo_id}/workers/stop":{"post":{"summary":"Stop Working On Wo","operationId":"stop_working_on_wo_work_orders__wo_id__workers_stop_post","parameters":[{"name":"wo_id","in":"path","required":true,"schema":{"type":"integer","title":"Wo Id"}},{"name":"authorization","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AddNoteRequest":{"properties":{"text":{"type":"string","maxLength":4000,"minLength":1,"title":"Text"}},"type":"object","required":["text"],"title":"AddNoteRequest"},"CloseWOResponse":{"properties":{"ok":{"type":"boolean","title":"Ok"},"status":{"type":"string","title":"Status"}},"type":"object","required":["ok","status"],"title":"CloseWOResponse"},"CreateUserRequest":{"properties":{"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"},"pin":{"type":"string","maxLength":6,"minLength":4,"title":"Pin"}},"type":"object","required":["name","role","pin"],"title":"CreateUserRequest"},"CreateWORequest":{"properties":{"station":{"type":"string","title":"Station"},"part_number":{"type":"string","title":"Part Number"},"customer_order":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Order"},"is_stock":{"type":"boolean","title":"Is Stock","default":false}},"type":"object","required":["station","part_number"],"title":"CreateWORequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"name":{"type":"string","title":"Name"},"pin":{"type":"string","maxLength":6,"minLength":4,"title":"Pin"}},"type":"object","required":["name","pin"],"title":"LoginRequest"},"LoginResponse":{"properties":{"token":{"type":"string","title":"Token"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"}},"type":"object","required":["token","name","role"],"title":"LoginResponse"},"NoteOut":{"properties":{"id":{"type":"integer","title":"Id"},"work_order_id":{"type":"integer","title":"Work Order Id"},"author_name":{"type":"string","title":"Author Name"},"station":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Station"},"text":{"type":"string","title":"Text"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","work_order_id","author_name","station","text","created_at"],"title":"NoteOut"},"ResetUserPinRequest":{"properties":{"user_id":{"type":"integer","title":"User Id"},"new_pin":{"type":"string","maxLength":6,"minLength":4,"title":"New Pin"}},"type":"object","required":["user_id","new_pin"],"title":"ResetUserPinRequest"},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","name","role","is_active"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WOOut":{"properties":{"id":{"type":"integer","title":"Id"},"wo_number":{"type":"string","title":"Wo Number"},"station":{"type":"string","title":"Station"},"part_number":{"type":"string","title":"Part Number"},"customer_order":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Customer Order"},"is_stock":{"type":"boolean","title":"Is Stock"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","wo_number","station","part_number","customer_order","is_stock","status","created_at"],"title":"WOOut"},"WorkerOut":{"properties":{"user_id":{"type":"integer","title":"User Id"},"name":{"type":"string","title":"Name"},"role":{"type":"string","title":"Role"},"started_at":{"type":"string","format":"date-time","title":"Started At"}},"type":"object","required":["user_id","name","role","started_at"],"title":"WorkerOut"}}}}